+++ src/appdistribution/types.ts
@@ -116,6 +116,17 @@ export interface ReleaseTest {
   deviceExecutions: DeviceExecution[];
   loginCredential?: LoginCredential;
   testCase?: string;
+  aiInstructions?: AIInstruction;
+}
+
+export interface AIInstruction {
+  steps: AIStep[];
+}
+
+export interface AIStep {
+  goal: string;
+  hint?: string;
+  successCriteria?: string;
 }
 
 export interface AiStep {