+++ src/deploy/dataconnect/prepare.ts
@@ -11,12 +11,11 @@ import { ensureApis } from "../../dataconnect/ensureApis";
 import { requireTosAcceptance } from "../../requireTosAcceptance";
 import { DATA_CONNECT_TOS_ID } from "../../gcp/firedata";
 import { setupCloudSql } from "../../dataconnect/provisionCloudSql";
-import { checkBillingEnabled } from "../../gcp/cloudbilling";
 import { parseServiceName } from "../../dataconnect/names";
 import { FirebaseError } from "../../error";
-import { requiresVector } from "../../dataconnect/types";
+import { mainSchema, requiresVector } from "../../dataconnect/types";
 import { diffSchema } from "../../dataconnect/schemaMigration";
-import { upgradeInstructions } from "../../dataconnect/freeTrial";
+import { checkBillingEnabled } from "../../gcp/cloudbilling";
 import { Context, initDeployStats } from "./context";
 
 /**
@@ -35,7 +34,6 @@ export default async function (context: Context, options: DeployOptions): Promis
   const { serviceInfos, filters, deployStats } = context.dataconnect;
   if (!(await checkBillingEnabled(projectId))) {
     deployStats.missingBilling = true;
-    throw new FirebaseError(upgradeInstructions(projectId));
   }
   await requireTosAcceptance(DATA_CONNECT_TOS_ID)(options);
   for (const si of serviceInfos) {
@@ -65,7 +63,7 @@ export default async function (context: Context, options: DeployOptions): Promis
     for (const si of serviceInfos) {
       await diffSchema(
         options,
-        si.schema,
+        mainSchema(si.schemas),
         si.dataConnectYaml.schema?.datasource?.postgresql?.schemaValidation,
       );
     }
@@ -76,7 +74,7 @@ export default async function (context: Context, options: DeployOptions): Promis
           return !filters || filters?.some((f) => si.dataConnectYaml.serviceId === f.serviceId);
         })
         .map(async (s) => {
-          const postgresDatasource = s.schema.datasources.find((d) => d.postgresql);
+          const postgresDatasource = mainSchema(s.schemas).datasources.find((d) => d.postgresql);
           if (postgresDatasource) {
             const instanceId = postgresDatasource.postgresql?.cloudSql?.instance.split("/").pop();
             const databaseId = postgresDatasource.postgresql?.database;